Replace backslash with left brace on base layer
Neither of those is very useful in an English layout, given that both
keys are more conveniently located on the symbols layer, but in the
German layout, the left brace keycode corresponds to Ü, which is quite
useful when typing German.  With this change, the umlaut keys are now in
their default position on the base layer with a German layout.